closes #81
